결제 api작성, 결제정보 레디스저장, 결제 시간제한

+++ order-service/src/main/java/com/example/orderservice/client/ProductServiceClient.java
@@ -16,9 +16,9 @@ public interface ProductServiceClient {
     ProductResponseDto findByIdStatusProduct(@PathVariable("productId") Long id);
 
     @PostMapping("/api/client/products/{id}/decrease-stock")
-    void decreaseStock(@PathVariable("id") Long productId, @RequestParam("count") Long count);
+    void decreaseStock(@PathVariable("id") Long productId, @RequestParam("count") int count);
 
     @PostMapping("/api/client/products/{id}/increase-stock")
-    void increaseStock(@PathVariable("id") Long productId, @RequestParam("count") Long count);
+    void increaseStock(@PathVariable("id") Long productId, @RequestParam("count") int count);
 
 }
